requirements.txt
google-api-python-client>=2.164.0
google-auth>=2.29.0
openpyxl>=3.1.2
pydantic>=2.8.2
PyYAML>=6.0.1
pytest>=8.2.0
python-dotenv>=1.0.1
Telethon>=1.36.0
